+++ src/main.cpp
@@ -191,8 +191,10 @@ int dispatch(Settings settings, const std::vector<std::string>& args) {
   // hidden commands (only in development versions)
   else if (cmd == "test") {
     commands.test();
-  } else if (cmd == "test-ie") {
+  } else if (cmd == "test-inceval") {
     commands.testIncEval();
+  } else if (cmd == "test-logeval") {
+    commands.testLogEval();
   } else if (cmd == "test-pari") {
     commands.testPari();
   } else if (cmd == "dot") {