+++ CHANGELOG.md
@@ -2,6 +2,18 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+## v22.12.14
+
+### Bugfixes
+
+* Fix validation for IE programs
+* Use transitive hash values for programs with `seq`
+
+### Enhancements
+
+* Add detection of programs with logarithmic complexity
+* Add stats for incremental and logarithmic eval
+
 ## v22.12.9
 
 ### Bugfixes